modification: moved sidebar code from dashboard to its own component file
